Web API Layer

+++ Entities/DTOs/CarDetailDto.cs
@@ -10,7 +10,7 @@ namespace Entities.DTOs
         public string CarName { get; set; }
         public string ColorName { get; set; }
         public string BrandName { get; set; }
-        public short DailyPrice { get; set; }
+        public decimal DailyPrice { get; set; }
 
     }
 }
